detect Linux distrobution without lsb_release
In order to detect the host's linux distribution,
Barreleye checks linux kernel revision by uname -r when
the lsb_release command is unavailable on the host.

If the program runs inside a container and this container's
distribution is different from that of its host, the build
program misdetects its distribution. (e.g. Builds inside
CentOS 7 container on Ubuntu host will be failed)

This patch uses /etc/redhat-release to detect the dirstribution.

Signed-off-by: envzhu <densikousakusyosinsya@gmail.com>
Signed-off-by: Li Xi <pkuelelixi@163.com>